// projeto/web/angular_p2/src/api/generated/models/responseUserDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Pizzaria API
 * API para gerenciamento de pedidos e clientes.
 * OpenAPI spec version: 1.0
 */
import type { EUserStatus } from './eUserStatus';
import type { ResponseRoleDto } from './responseRoleDto';

export interface ResponseUserDto {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  id?: number | string;
  email?: string;
  name?: string;
  status?: EUserStatus;
  /**
   * @nullable
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  roleId?: number | string | null;
  role?: null | ResponseRoleDto;
}